Making the Equipped Backpack available as a Container for purposes of Crafting Addons
- While this probably won't pick up as a Container the way that most of the mods detect containers, this does allow those mods an opportunity to see if there's a container on the "Player" object, and if so, add it as a container.
